+++ internal/ui/sections/volumes/section.go
@@ -172,7 +172,7 @@ func (s *Section) Update(msg tea.Msg) tea.Cmd {
 		summary := fmt.Sprintf(
 			"Pruned %d volumes, reclaimed %s",
 			msg.report.ItemsDeleted,
-			helper.FormatSize(int64(msg.report.SpaceReclaimed)),
+			helper.FormatSize(msg.report.SpaceReclaimed),
 		)
 		return tea.Batch(s.updateVolumesCmd(), func() tea.Msg {
 			return message.ShowBannerMsg{Message: summary, IsError: false}
